--- index.js
@@ -1,18 +1,0 @@
-module.exports =
-{
-	SDPInfo			: require("./lib/SDPInfo"),
-	CandidateInfo		: require("./lib/CandidateInfo"),
-	CodecInfo		: require("./lib/CodecInfo"),
-	DTLSInfo		: require("./lib/DTLSInfo"),
-	CryptoInfo		: require("./lib/CryptoInfo"),
-	ICEInfo			: require("./lib/ICEInfo"),
-	MediaInfo		: require("./lib/MediaInfo"),
-	Setup			: require("./lib/Setup"),
-	SourceGroupInfo		: require("./lib/SourceGroupInfo"),
-	SourceInfo		: require("./lib/SourceInfo"),
-	StreamInfo		: require("./lib/StreamInfo"),
-	TrackInfo		: require("./lib/TrackInfo"),
-	RTCPFeedbackInfo       	: require("./lib/RTCPFeedbackInfo"),
-	TrackEncodingInfo       : require("./lib/TrackEncodingInfo"),
-	Direction		: require("./lib/Direction")
-};
